examples/python-flask-smtp/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.8.0
python-dotenv==1.0.1
